requirements.txt
accelerate==0.34.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
async-timeout==4.0.3
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.0.post1
charset-normalizer==3.4.6
click==8.1.8
cmdstanpy==1.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.6
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.20
decorator==5.2.1
Deprecated==1.3.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.135.2
ffmpy==1.0.0
filelock==3.25.2
finance-datareader==0.9.96
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.46
gradio==6.10.0
gradio_client==2.4.0
greenlet==3.3.2
groovy==0.1.2
gTTS==2.5.4
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.2
holidays==0.93
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
ipykernel==7.2.0
ipython==8.38.0
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.83
langchain-mcp-adapters==0.2.0
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.15
langsmith==0.7.22
limits==3.13.0
llvmlite==0.46.0
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.9.4
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.16
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.64.0
numpy==2.2.6
openai==2.30.0
openai-whisper @ https://github.com/openai/whisper/archive/refs/heads/main.zip#sha256=028271fff5d90da72111b58aa3b1bc3092ff66241fa76ee11f5ea96f63e3a632
opencv-python==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
parso==0.8.6
peewee==4.0.3
pillow==11.3.0
platformdirs==4.9.4
plotly==6.0.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.1.6
protobuf==6.33.6
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
redis==5.0.4
referencing==0.37.0
regex==2026.2.28
reportlab==4.2.5
requests==2.33.0
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.3.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sse-starlette==3.3.3
stack-data==0.6.3
stanio==0.5.1
starlette==0.52.1
streamlit==1.55.0
sympy==1.13.1
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==4.44.0
typer==0.23.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchdog==6.0.0
wcwidth==0.6.0
webencodings==0.5.1
widgetsnbextension==4.0.15
wrapt==2.1.2
xgboost==3.2.0
xxhash==3.6.0
yarl==1.23.0
yfinance==0.2.52
zstandard==0.25.0
